+++ HomeGroup.API/Models/DTOs/Attendance/AttendanceDtos.cs
@@ -1,9 +1,9 @@
 namespace HomeGroup.API.Models.DTOs.Attendance;
 
-public record RecordAttendanceRequest(int HomeGroupId, DateOnly MeetingDate, List<AttendanceEntry> Entries);
+public record RecordAttendanceRequest(long HomeGroupId, DateOnly MeetingDate, List<AttendanceEntry> Entries);
 
-public record AttendanceEntry(int PersonId, bool WasPresent, string? Notes);
+public record AttendanceEntry(long PersonId, bool WasPresent, string? Notes);
 
-public record AttendanceResponse(int Id, int PersonId, string PersonName, int HomeGroupId, DateOnly MeetingDate, bool WasPresent, string? Notes);
+public record AttendanceResponse(long Id, long PersonId, string PersonName, long HomeGroupId, DateOnly MeetingDate, bool WasPresent, string? Notes);
 
 public record AttendanceSummary(DateOnly MeetingDate, int TotalMembers, int PresentCount, double AttendanceRate);
