+++ HomeGroup.API/Models/DTOs/Groups/GroupDtos.cs
@@ -1,9 +1,9 @@
 namespace HomeGroup.API.Models.DTOs.Groups;
 
-public record CreateGroupRequest(string Name, string? Description, string? MeetingDay, string? MeetingTime, string? Location, int? LeaderId);
+public record CreateGroupRequest(string Name, string? Description, string? MeetingDay, string? MeetingTime, string? Location, long? LeaderId);
 
-public record UpdateGroupRequest(string Name, string? Description, string? MeetingDay, string? MeetingTime, string? Location, int? LeaderId, bool IsActive);
+public record UpdateGroupRequest(string Name, string? Description, string? MeetingDay, string? MeetingTime, string? Location, long? LeaderId, bool IsActive);
 
-public record GroupResponse(int Id, string Name, string? Description, string? MeetingDay, string? MeetingTime, string? Location, int? LeaderId, string? LeaderName, bool IsActive, int MemberCount);
+public record GroupResponse(long Id, string Name, string? Description, string? MeetingDay, string? MeetingTime, string? Location, long? LeaderId, string? LeaderName, bool IsActive, int MemberCount);
 
-public record AddMemberRequest(int PersonId, string Role = "Member");
+public record AddMemberRequest(long PersonId, string Role = "Member");
