+++ HomeGroup.API/Models/Entities/Attendance.cs
@@ -2,10 +2,10 @@ namespace HomeGroup.API.Models.Entities;
 
 public class Attendance
 {
-    public int Id { get; set; }
-    public int PersonId { get; set; }
+    public long Id { get; set; }
+    public long PersonId { get; set; }
     public Person Person { get; set; } = null!;
-    public int HomeGroupId { get; set; }
+    public long HomeGroupId { get; set; }
     public HomeGroupEntity HomeGroup { get; set; } = null!;
     public DateOnly MeetingDate { get; set; }
     public bool WasPresent { get; set; }
